pyproject.toml
[tool.poetry]
name = "thaitech-solutions"
version = "0.1.0"
description = "Professional web development solutions for Thailand"
authors = ["ThaiTech Solutions Team <contact@thaitechsolutions.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
flask-babel = "^2.0.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
